+++ TradingSystem611/src/main/java/org/TradingSystem/model/Tradeable.java
@@ -1,6 +1,6 @@
 package org.TradingSystem.model;
 
 public interface Tradeable {
-    int getPrice();
-    void changePrice(int price);
+    double getPrice();
+    void changePrice(double price);
 }
